pyproject.toml
[project]
name = "todo"
version = "0.1.0"
description = "ToDo bot"
requires-python = ">=3.13"
dependencies = [
    "pydantic-settings>=2.12.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.9",
    "pre-commit>=4.5.1",
]

backend = [
    "celery>=5.6.0",
    "django>=6.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "djangorestframework-stubs>=3.16.8",
    "gunicorn>=23.0.0",
    "psycopg[binary]>=3.3.2",
    "adrf>=0.1.12",
    "asyncpg>=0.31.0",
    "daphne>=4.2.1",
    "drf-spectacular>=0.29.0",
]

bot = [
    "aiogram>=3.23.0",
    "aiogram-dialog>=2.4.0",
    "redis>=7.1.0",
]

[tool.ruff]
target-version = "py313"
line-length = 88
exclude = [
    ".git",
    "__pycache__",
    "migrations",
    "venv",
    ".venv",
]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "B",
    "DJ",
    "UP",
    "TCH",
]
ignore = ["E501"]
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.isort]
known-first-party = ["backend", "bot"]
combine-as-imports = true
